Update flash.py
The default scan range for com ports is set to 1-257. since there are only 256 available ports, it defaults back to the default public set of com ports 1-8. This caused a problem since a large majority of the devices register at 13 or higher by default upon initial connection. By changing line 112 from (1,257) to f(1,256) it now scans all available ports for the user